+++ src/main/ts/BaseToolSidebar.tsx
@@ -1,0 +1,12 @@
+namespace contoureditor {
+
+    interface BaseToolSidebarProps {
+        isActive: boolean
+    }
+
+    export class BaseToolSidebar extends React.Component<BaseToolSidebarProps,{}>{
+        public static defaultProps = {
+            isActive: false
+        }
+    }
+}