+++ src/main/ts/Sidebar.tsx
@@ -1,9 +1,23 @@
+///<reference path="BaseToolSidebar.tsx"/>
+///<reference path="CreateToolSidebar.tsx"/>
+///<reference path="SelectToolSidebar.tsx"/>
+///<reference path="MoveToolSidebar.tsx"/>
+
 namespace contoureditor {
-    export class Sidebar extends React.Component {
+
+    interface SidebarProps {
+        tool: string
+    }
+
+    export class Sidebar extends React.Component<SidebarProps,{}> {
 
         public render() {
             return (
-                <div id="sidebar">sidebar</div>
+                <div id="sidebar">
+                    <CreateToolSidebar isActive={this.props.tool == "create"}/>
+                    <SelectToolSidebar isActive={this.props.tool == "select"}/>
+                    <MoveToolSidebar isActive={this.props.tool == "move"}/>
+                </div>
             );
         }
     }